Use `SimulateSystem` for `uses_from_macos` deps
