+++ Library/Homebrew/test/software_spec_spec.rb
@@ -135,27 +135,72 @@ describe SoftwareSpec do
     end
   end
 
-  describe "#uses_from_macos" do
-    it "allows specifying dependencies", :needs_linux do
-      spec.uses_from_macos("foo")
+  describe "#uses_from_macos", :needs_linux do
+    context "when running on Linux", :needs_linux do
+      it "allows specifying dependencies" do
+        spec.uses_from_macos("foo")
 
-      expect(spec.deps.first.name).to eq("foo")
+        expect(spec.deps.first.name).to eq("foo")
+      end
+
+      it "works with tags" do
+        spec.uses_from_macos("foo" => :build)
+
+        expect(spec.deps.first.name).to eq("foo")
+        expect(spec.deps.first.tags).to include(:build)
+      end
+
+      it "ignores OS version specifications" do
+        spec.uses_from_macos("foo", since: :mojave)
+        spec.uses_from_macos("bar" => :build, :since => :mojave)
+
+        expect(spec.deps.first.name).to eq("foo")
+        expect(spec.deps.last.name).to eq("bar")
+        expect(spec.deps.last.tags).to include(:build)
+      end
     end
 
-    it "works with tags", :needs_linux do
-      spec.uses_from_macos("foo" => :build)
+    context "when running on macOS", :needs_macos do
+      before do
+        allow(OS).to receive(:mac?).and_return(true)
+        allow(OS::Mac).to receive(:version).and_return(OS::Mac::Version.from_symbol(:sierra))
+      end
 
-      expect(spec.deps.first.name).to eq("foo")
-      expect(spec.deps.first.tags).to include(:build)
-    end
+      it "adds a macOS dependency if the OS version meets requirements" do
+        spec.uses_from_macos("foo", since: :el_capitan)
 
-    it "ignores OS version specifications", :needs_linux do
-      spec.uses_from_macos("foo", since: :mojave)
-      spec.uses_from_macos("bar" => :build, :since => :mojave)
+        expect(spec.deps).to be_empty
+        expect(spec.uses_from_macos_elements.first).to eq("foo")
+      end
 
-      expect(spec.deps.first.name).to eq("foo")
-      expect(spec.deps.last.name).to eq("bar")
-      expect(spec.deps.last.tags).to include(:build)
+      it "doesn't add a macOS dependency if the OS version doesn't meet requirements" do
+        spec.uses_from_macos("foo", since: :high_sierra)
+
+        expect(spec.deps.first.name).to eq("foo")
+        expect(spec.uses_from_macos_elements).to eq(["foo"])
+      end
+
+      it "works with tags" do
+        spec.uses_from_macos("foo" => :build, :since => :high_sierra)
+
+        dep = spec.deps.first
+
+        expect(dep.name).to eq("foo")
+        expect(dep.tags).to include(:build)
+      end
+
+      it "doesn't add a dependency if no OS version is specified" do
+        spec.uses_from_macos("foo")
+        spec.uses_from_macos("bar" => :build)
+
+        expect(spec.deps).to be_empty
+      end
+
+      it "raises an error if passing invalid OS versions" do
+        expect {
+          spec.uses_from_macos("foo", since: :bar)
+        }.to raise_error(MacOSVersionError, "unknown or unsupported macOS version: :bar")
+      end
     end
   end
